Split main() and runAuthTest() into a separate class.
