+++ SCMPToNVPTest/ics_util.cpp
@@ -13,20 +13,15 @@
 #include <codecvt>
 #include <fstream>
 #include <sstream>
-#include "ics.h"
 #include "soapINVPTransactionProcessorProxy.h"
 #include "NVPCybersource.h"
+#include "ics.h"
+#include "ics_util.h"
 
-#define SCMP_REQUEST_ICS_APPLICATIONS "SCMP_REQUEST_ICS_APPLICATIONS"
-#define MAX_KEY_LENGTH 100
-#define MAX_VALUE_LENGTH 100
 
-const char CYBS_INI_FILE[]   = "../resources/cybs.ini";
 
 void printMap(std::map <std::wstring, std::wstring> m);
 std::map<std::wstring, std::wstring> loadPropertiesFile(const std::string& filename);
-void runAuthTest();
-ics_msg* processRequest(ics_msg* icsRequest);
 std::map <std::wstring, std::wstring> convertICSRequestToSimpleOrderRequest(ics_msg* icsRequest);
 ics_msg* convertSimpleOrderResponseToICSResponse(std::map <std::wstring, std::wstring> soResponse);
 char* wstring_to_char(const std::wstring& wstr);
@@ -35,61 +30,6 @@ std::vector<std::wstring> splitWString(const std::wstring& str, wchar_t delimite
 std::wstring stringToWString(const std::string& str);
 std::string wstringToString(const std::wstring& wstr);
 
-int main() {
-    printf("Running auth transaction\n");
-    runAuthTest();
-    return 0;
-}
-
-void printMap(std::map <std::wstring, std::wstring> m) {
-    typedef std::map <std::wstring, std::wstring>::const_iterator it_type;
-    for (it_type iterator = m.begin(); iterator != m.end(); iterator++) {
-#ifdef WIN32
-        wprintf(L"%s", iterator->first.c_str());
-        wprintf(L"==>");
-        wprintf(L"%s", iterator->second.c_str());
-        wprintf(L"\n");
-#else
-        wprintf(L"%S", iterator->first.c_str());
-        wprintf(L"==>");
-        wprintf(L"%S", iterator->second.c_str());
-        wprintf(L"\n");
-#endif
-
-    }
-}
-void runAuthTest() {
-    float amount = 20.00;
-    char offer[1024];
-
-    ics_msg* icsorder;
-    icsorder = ics_init(0);
-
-    ics_fadd(icsorder, (char*)"ics_applications", (char*)"ics_auth");
-    ics_fadd(icsorder, (char*)"customer_firstname", (char*)"John");
-    ics_fadd(icsorder, (char*)"customer_lastname", (char*)"Doe");
-    ics_fadd(icsorder, (char*)"customer_email", (char*)"nobody@cybersource.com");
-    ics_fadd(icsorder, (char*)"customer_phone", (char*)"408-556-9100");
-    ics_fadd(icsorder, (char*)"bill_address1", (char*)"1295 Charleston Rd.");
-    ics_fadd(icsorder, (char*)"bill_city", (char*)"Mountain View");
-    ics_fadd(icsorder, (char*)"bill_state", (char*)"CA");
-    ics_fadd(icsorder, (char*)"bill_zip", (char*)"94043-1307");
-    ics_fadd(icsorder, (char*)"bill_country", (char*)"US");
-    ics_fadd(icsorder, (char*)"customer_cc_number", (char*)"4111111111111111");
-    ics_fadd(icsorder, (char*)"customer_cc_expmo", (char*)"12");
-    ics_fadd(icsorder, (char*)"customer_cc_expyr", (char*)"2030");
-    ics_fadd(icsorder, (char*)"merchant_ref_number", (char*)"12");
-    ics_fadd(icsorder, (char*)"currency", (char*)"USD");
-
-    //add 1 item
-    sprintf(offer, "amount:%.2f^merchant_product_sku:GC1^product_name:Gift Certificate^quantity:1",
-        amount);
-    ics_fadd(icsorder, (char*)"offer0", offer);
-
-    ics_msg* icsResponse = processRequest(icsorder);
-    ics_destroy(icsorder);
-    ics_destroy(icsResponse);
-}
 
 /**
  * Process the given ICS request by translating it to a Simple Order request and send to the SO API. The SO response is then converted to ICS response object.
@@ -110,33 +50,6 @@ ics_msg* processRequest(ics_msg* icsRequest) {
     soap_set_imode(proxy.soap, SOAP_C_MBSTRING);
 
     std::map <std::wstring, std::wstring> request;
-
-    /* Original SO request
-    request[L"merchantReferenceCode"] = L"your_merchant_reference_code";
-    request[L"billTo_firstName"] = L"AJŠAß";
-    request[L"billTo_lastName"] = L"Doe";
-    request[L"billTo_street1"] = L"1295 Charleston Road";
-    request[L"billTo_city"] = L"Mountain View";
-    request[L"billTo_state"] = L"CA";
-    request[L"billTo_postalCode"] = L"94043";
-    request[L"billTo_country"] = L"US";
-    request[L"billTo_email"] = L"nobody@cybersource.com";
-    request[L"billTo_ipAddress"] = L"10.7.7.7";
-    request[L"billTo_phoneNumber"] = L"650-965-6000";
-    request[L"shipTo_firstName"] = L"Jane";
-    request[L"shipTo_lastName"] = L"Jane";
-    request[L"shipTo_street1"] = L"123 toile délavée noir";
-    request[L"shipTo_state"] = L"CA";
-    request[L"shipTo_city"] = L"CA";
-    request[L"shipTo_postalCode"] = L"94401";
-    request[L"shipTo_country"] = L"US";
-    request[L"card_accountNumber"] = L"xxxxxxxxxxxxxxxx";
-    request[L"card_expirationMonth"] = L"12";
-    request[L"card_expirationYear"] = L"2025";
-    request[L"purchaseTotals_currency"] = L"USD";
-    request[L"item_0_unitPrice"] = L"12.34";
-    request[L"ccAuthService_run"] = L"true";
-    */
 
     request = convertICSRequestToSimpleOrderRequest(icsRequest);
     printf("\n==== Simple Order Request ====\n");
@@ -172,33 +85,26 @@ ics_msg* processRequest(ics_msg* icsRequest) {
     return icsResponse;
 }
 
+void printMap(std::map <std::wstring, std::wstring> m) {
+    typedef std::map <std::wstring, std::wstring>::const_iterator it_type;
+    for (it_type iterator = m.begin(); iterator != m.end(); iterator++) {
+#ifdef WIN32
+        wprintf(L"%s", iterator->first.c_str());
+        wprintf(L"==>");
+        wprintf(L"%s", iterator->second.c_str());
+        wprintf(L"\n");
+#else
+        wprintf(L"%S", iterator->first.c_str());
+        wprintf(L"==>");
+        wprintf(L"%S", iterator->second.c_str());
+        wprintf(L"\n");
+#endif
+
+    }
+}
+
 std::map <std::wstring, std::wstring> convertICSRequestToSimpleOrderRequest(ics_msg* icsRequest) {
     std::map <std::wstring, std::wstring> soRequest;
-
-    /* hard-coded static request
-    soRequest[L"merchantReferenceCode"] =  charToWString(ics_fgetbyname(icsRequest, "merchant_ref_number"));
-    soRequest[L"billTo_firstName"] = charToWString(ics_fgetbyname(icsRequest, "customer_firstname"));
-    soRequest[L"billTo_lastName"] = charToWString(ics_fgetbyname(icsRequest, "customer_lastname"));
-    soRequest[L"billTo_street1"] = charToWString(ics_fgetbyname(icsRequest, "bill_address1"));
-    soRequest[L"billTo_city"] = charToWString(ics_fgetbyname(icsRequest, "bill_city"));
-    soRequest[L"billTo_state"] = charToWString(ics_fgetbyname(icsRequest, "bill_state"));
-    soRequest[L"billTo_postalCode"] = charToWString(ics_fgetbyname(icsRequest, "bill_zip"));
-    soRequest[L"billTo_country"] = charToWString(ics_fgetbyname(icsRequest, "bill_country"));
-    soRequest[L"billTo_email"] = charToWString(ics_fgetbyname(icsRequest, "customer_email"));
-    soRequest[L"billTo_ipAddress"] = charToWString(ics_fgetbyname(icsRequest, "customer_ipaddress"));
-    soRequest[L"billTo_phoneNumber"] = charToWString(ics_fgetbyname(icsRequest, "customer_phone"));
-    soRequest[L"shipTo_firstName"] = charToWString(ics_fgetbyname(icsRequest, "ship_to_firstname"));
-    soRequest[L"shipTo_lastName"] = charToWString(ics_fgetbyname(icsRequest, "ship_to_lastname"));
-    soRequest[L"shipTo_street1"] = charToWString(ics_fgetbyname(icsRequest, "ship_to_address1"));
-    soRequest[L"shipTo_state"] = charToWString(ics_fgetbyname(icsRequest, "ship_to_state"));
-    soRequest[L"shipTo_city"] = charToWString(ics_fgetbyname(icsRequest, "ship_to_city"));
-    soRequest[L"shipTo_postalCode"] = charToWString(ics_fgetbyname(icsRequest, "ship_to_zip"));
-    soRequest[L"shipTo_country"] = charToWString(ics_fgetbyname(icsRequest, "ship_to_country"));
-    soRequest[L"card_accountNumber"] = charToWString(ics_fgetbyname(icsRequest, "customer_cc_number"));
-    soRequest[L"card_expirationMonth"] = charToWString(ics_fgetbyname(icsRequest, "customer_cc_expmo"));
-    soRequest[L"card_expirationYear"] = charToWString(ics_fgetbyname(icsRequest, "customer_cc_expyr"));
-    soRequest[L"purchaseTotals_currency"] = charToWString(ics_fgetbyname(icsRequest, "currency"));
-    */
 
     // loads the request mapping table
     std::map<std::wstring, std::wstring> requestMap = loadPropertiesFile("scmp_so_mapping.properties");
@@ -289,11 +195,6 @@ std::map <std::wstring, std::wstring> convertICSRequestToSimpleOrderRequest(ics_
         }
     }
 
-    //TODO: parse the item/offer text here
-    //soRequest[L"item_0_unitPrice"] = L"4.59";
-
-    //soRequest[L"ccAuthService_run"] = L"true";
-
     return soRequest;
 }
 
@@ -301,39 +202,6 @@ ics_msg* convertSimpleOrderResponseToICSResponse(std::map <std::wstring, std::ws
     ics_msg* icsResponse;
     icsResponse = ics_init(0);
 
-    /*
-    Sample Simple Order response for reference
-
-    purchaseTotals_currency=USD
-    ccAuthReply_cardPINlessDebit=Y
-    ccAuthReply_authorizationCode=888888
-    ccAuthReply_cardRegulated=Y
-    ccAuthReply_cardPayroll=Y
-    ccAuthReply_cardCommercial=Y
-    card_cardType=001
-    ccAuthReply_cardIssuerCountry=840
-    ccAuthReply_cardHealthcare=Y
-    ccAuthReply_cardLevel3Eligible=Y
-    decisionEarlyReply_rcode=1
-    ccAuthReply_cardSignatureDebit=Y
-    reasonCode=100
-    ccAuthReply_cardPrepaid=Y
-    ccAuthReply_processorResponse=100
-    ccAuthReply_paymentNetworkTransactionID=123456789619999
-    pos_terminalID=123456
-    ccAuthReply_avsCode=X
-    decision=ACCEPT
-    ccAuthReply_reasonCode=100
-    merchantReferenceCode=080
-    ccAuthReply_reconciliationID=70877315D5XP6CL7
-    ccAuthReply_amount=42.97
-    requestID=7376721305846381904602
-    ccAuthReply_authorizedDateTime=2025-01-23T22:42:11Z
-    requestToken=Axj/7wSTj+SSsB9tQYLaABEg3YOG7dmxaxGtig2hzG6e/IAOSTAKe/IAOSTaQNhkpz8MmkmXoxh9qbwJycfySVgPtqDBbQAA8TJd
-    ccAuthReply_affluenceIndicator=Y
-    ccAuthReply_avsCodeRaw=I1
-    decisionEarlyReply_reasonCode=100
-    */
     std::map<std::wstring, std::wstring> responseMap = loadPropertiesFile("so_scmp_response_mapping.properties");
 
     for (const auto& pair : soResponse) {
@@ -377,26 +245,6 @@ char* wstring_to_char(const std::wstring& wstr) {
 
     return char_array;
 }
-
-/*
-void printMap (std::map <std::wstring, std::wstring> m) {
-    typedef std::map <std::wstring, std::wstring>::const_iterator it_type;
-    for(it_type iterator = m.begin(); iterator != m.end(); iterator++) {
-        #ifdef WIN32
-            wprintf (L"%s", iterator->first.c_str());
-            wprintf (L"==>");
-            wprintf (L"%s",iterator->second.c_str());
-            wprintf (L"\n");
-        #else
-            wprintf (L"%S", iterator->first.c_str());
-                        wprintf (L"==>");
-                        wprintf (L"%S",iterator->second.c_str());
-                        wprintf (L"\n");
-        #endif
-
-    }
-}
-*/
 
 std::map<std::wstring, std::wstring> loadPropertiesFile(const std::string& filename) {
     std::map<std::wstring, std::wstring> properties;
